MGS-728 Remove use of assemble and use handlebars directly.
This consolidates the code used for generating the static and the code
for generating previews.

Previously, each site used its own grunt-assemble configuration to
generate the site, and site-builder used its version of assemble to
generate previews. Consolidating ensures that static site generation and
preview use the same version of handlebars.